Split sql-plugin override registries
