+++ sql-plugin/src/main/scala/org/apache/spark/sql/rapids/aggregate/aggregateBase.scala
@@ -279,6 +279,39 @@ case class GpuAggregateExpression(origAggregateFunction: GpuAggregateFunction,
   override def sql: String = aggregateFunction.sql(isDistinct)
 }
 
+case class GpuAggregateExpressionMeta(
+    aggregateExpression: AggregateExpression,
+    override val conf: RapidsConf,
+    parentMeta: Option[RapidsMeta[_, _, _]],
+    rule: DataFromReplacementRule)
+    extends ExprMeta[AggregateExpression](aggregateExpression, conf, parentMeta, rule) {
+
+  private val filter: Option[BaseExprMeta[_]] =
+    aggregateExpression.filter.map(GpuOverrides.wrapExpr(_, conf, Some(this)))
+  private val childrenExprMeta: Seq[BaseExprMeta[Expression]] =
+    aggregateExpression.children.map(GpuOverrides.wrapExpr(_, conf, Some(this)))
+  override val childExprs: Seq[BaseExprMeta[_]] = childrenExprMeta ++ filter.toSeq
+
+  override def convertToGpuImpl(): GpuExpression = {
+    // Handle the case where AggregateExpression has the resultIds parameter containing a
+    // sequence of ExprIds instead of a single resultId.
+    val resultId = try {
+      val resultMethod = aggregateExpression.getClass.getMethod("resultId")
+      resultMethod.invoke(aggregateExpression).asInstanceOf[ExprId]
+    } catch {
+      case _: Exception =>
+        val resultMethod = aggregateExpression.getClass.getMethod("resultIds")
+        resultMethod.invoke(aggregateExpression).asInstanceOf[Seq[ExprId]].head
+    }
+    GpuAggregateExpression(
+      childExprs.head.convertToGpu().asInstanceOf[GpuAggregateFunction],
+      aggregateExpression.mode,
+      aggregateExpression.isDistinct,
+      filter.map(_.convertToGpu()),
+      resultId)
+  }
+}
+
 trait CudfAggregate extends Serializable {
   // we use this to get the ordinal of the bound reference, s.t. we can ask cudf to perform
   // the aggregate on that column
